Refactor user APIs according to 'authenticateJWT'

+++ src/apis/user.ts
@@ -3,7 +3,14 @@ import { PrismaClient } from '@prisma/client';
 const prisma = new PrismaClient();
 
 export const getUser = async (req, res) => {
-  const { userId } = req.params;
+  const userId = req.user?.id;
+  if (!userId) {
+    return res.status(400).send({
+      error: 'notLoggedIn',
+      message: 'This user is not logged in!',
+    });
+  }
+
   const user = await prisma.user.findUnique({
     where: {
       id: userId,
@@ -16,7 +23,14 @@ export const getUser = async (req, res) => {
 };
 
 export const updateUser = async (req, res) => {
-  const { userId } = req.params;
+  const userId = req.user?.id;
+  if (!userId) {
+    return res.status(400).send({
+      error: 'notLoggedIn',
+      message: 'This user is not logged in!',
+    });
+  }
+
   const { email, name } = req.body.data;
 
   const user = await prisma.user.update({
@@ -35,7 +49,13 @@ export const updateUser = async (req, res) => {
 };
 
 export const deleteUser = async (req, res) => {
-  const { userId } = req.params;
+  const userId = req.user?.id;
+  if (!userId) {
+    return res.status(400).send({
+      error: 'notLoggedIn',
+      message: 'This user is not logged in!',
+    });
+  }
 
   const user = await prisma.user.delete({
     where: {
